Compliance: refactor handlePackages duplicate
