+++ src/Components/CreateImageWizard/steps/Oscap/components/ProfileSelector.tsx
@@ -82,7 +82,7 @@ const ProfileSelector = () => {
   >([]);
   const complianceType = useAppSelector(selectComplianceType);
   const prefetchProfile = useBackendPrefetch('getOscapCustomizations');
-  const { clearCompliancePackages } = useSelectorHandlers();
+  const { clearCompliancePackages, handlePackages } = useSelectorHandlers();
 
   const {
     data: profiles,
@@ -191,23 +191,6 @@ const ProfileSelector = () => {
     setFilterValue('');
   };
 
-  const handlePackages = (
-    oldOscapPackages: string[],
-    newOscapPackages: string[]
-  ) => {
-    clearCompliancePackages(oldOscapPackages);
-
-    for (const pkg of newOscapPackages) {
-      dispatch(
-        addPackage({
-          name: pkg,
-          summary: 'Required by chosen OpenSCAP profile',
-          repository: 'distro',
-        })
-      );
-    }
-  };
-
   const handlePartitions = (oscapPartitions: Filesystem[]) => {
     dispatch(clearPartitions());
 
@@ -309,7 +292,11 @@ const ProfileSelector = () => {
           const oscapPartitions = response.filesystem || [];
           const newOscapPackages = response.packages || [];
           handlePartitions(oscapPartitions);
-          handlePackages(oldOscapPackages, newOscapPackages);
+          handlePackages(
+            oldOscapPackages,
+            newOscapPackages,
+            'Required by chosen OpenSCAP profile'
+          );
           handleServices(response.services);
           handleKernelAppend(response.kernel?.append);
           dispatch(
