+++ src/Components/CreateImageWizard/steps/Oscap/components/useSelectorHandlers.tsx
@@ -1,5 +1,5 @@
 import { useAppDispatch } from '../../../../../store/hooks';
-import { removePackage } from '../../../../../store/wizardSlice';
+import { addPackage, removePackage } from '../../../../../store/wizardSlice';
 
 export const useSelectorHandlers = () => {
   const dispatch = useAppDispatch();
@@ -10,7 +10,26 @@ export const useSelectorHandlers = () => {
     }
   };
 
+  const handlePackages = (
+    oldOscapPackages: string[],
+    newOscapPackages: string[],
+    reason: string
+  ) => {
+    clearCompliancePackages(oldOscapPackages);
+
+    for (const pkg of newOscapPackages) {
+      dispatch(
+        addPackage({
+          name: pkg,
+          summary: reason,
+          repository: 'distro',
+        })
+      );
+    }
+  };
+
   return {
     clearCompliancePackages,
+    handlePackages,
   };
 };
